+++ arviz/plots/forestplot.py
@@ -246,7 +246,7 @@ def plot_forest(
         width_ratios.append(1)
 
     if hdi_prob is None:
-        hdi_prob = rcParams["stats.hdi_prob"]
+        hdi_prob = rcParams["stats.ci_prob"]
     elif not 1 >= hdi_prob > 0:
         raise ValueError("The value of hdi_prob should be in the interval (0, 1]")