+++ arviz/plots/posteriorplot.py
@@ -237,7 +237,7 @@ def plot_posterior(
         labeller = BaseLabeller()
 
     if hdi_prob is None:
-        hdi_prob = rcParams["stats.hdi_prob"]
+        hdi_prob = rcParams["stats.ci_prob"]
     elif hdi_prob not in (None, "hide"):
         if not 1 >= hdi_prob > 0:
             raise ValueError("The value of hdi_prob should be in the interval (0, 1]")