+++ examples/matplotlib/mpl_plot_ecdf.py
@@ -4,6 +4,7 @@ ECDF Plot
 _gallery_category: Distributions
 """
 
+import warnings
 import matplotlib.pyplot as plt
 from scipy.stats import norm
 
@@ -13,6 +14,8 @@ az.style.use("arviz-doc")
 
 sample = norm(0, 1).rvs(1000)
 distribution = norm(0, 1)
+
+warnings.filterwarnings("ignore", category=az.utils.BehaviourChangeWarning)
 
 az.plot_ecdf(sample, cdf=distribution.cdf, confidence_bands=True)